refactor: movido enum permissões de projeto 'app' para 'api'
O Enum de permissões foi transferido para o arquivo Enums.cs já existente em api
